Completed basic templating of colorbars and run through rmse / fss multidate / multilead
Preparing to clean out old plotting templates for heat plots.  Need to integrate
remaining templates into the class heirarchy, currently finished main plot.
