+++ src/plotting/__init__.py
@@ -48,6 +48,7 @@ import matplotlib.pyplot as plt
 import seaborn as sns
 import numpy as np
 import pandas as pd
+import re
 import pickle
 import os
 from attrs import define, field, validators
